Scrub diagnostics in markdown in json test files
It repeats information in `derived.json`, just in a harder to inspect
way due to long line of text.  The regexp that is used is careful to
delete only these things (there are more uses of `<details>` for the
benchmark results and for many reviewers).

Done both in the testing code and in `create-fixture.ts`.  I also
removed it manually from existing input files to keep things compact.

(And also some more hand linting.)

Closes #127.