Implement a file fetcher, with the ability to mock it in tests
Currently used only for the owners.  Has a `limit` argument to avoid
getting whole index files since the owners are listed at the top in all
of them.  The limit that is used is 10k, and the current file with the
most maintainers has them listed in the first 3k.

Get's rid of `getOwnersOfPackages.ts`, which is just a function now.