Slight improvement for single owners
Make the main checklist show

> Most recent commit is approved by type definition owners or DT maintainers

It still is confusing that it's only a minor phrasing response for
blessing the PR, but it's still in the maintainer queue column.  It
might be better to have something more concrete for single owners, but
anything I can think about is a major change.