Ignore outdated "ready to merge" comments
Send a few dates to `usersSayReadyToMerge`, which uses them to check for
a comment with that content that was added *after* these dates.  So, for
example, if that comment was added and only later a review came in (that
makes auto-merging possible), then that comment will be ignored.  Note
that this does *not* include comment dates, so even if someone posted a
comment after the request (in the short time that the bot waits before
acting), it will still get merged.

Fixes #164.