Make path mapping edits okay
The ultimate goal here is to require owner, not maintainer approval for
path mapping edits, and the rationale is that there's already fairly
good coverage of path mappings in @definitelytyped/definitions-parser:
- <https://github.com/microsoft/DefinitelyTyped-tools/blob/314aebafbcc713b11fe06d9ba34e334b1ec303e1/packages/definitions-parser/src/lib/definition-parser.ts#L408>
- <https://github.com/microsoft/DefinitelyTyped-tools/blob/314aebafbcc713b11fe06d9ba34e334b1ec303e1/packages/definitions-parser/src/check-parse-results.ts#L82>
- microsoft/DefinitelyTyped-tools#120

Off the top of my head, about the only not-strictly-necessary edits that
could get by the CI are:
- Extra, unneeded scope path mappings (`"@notinuse/*": ["notinuse__*"]`)
- Extra, unneeded wildcard path mappings (`"react-dom/*":
  ["react-dom/v15/*"]`)
- That's it?

So if the CI has passed, I'm not sure what additional scrutiny a
maintainer offers. I think it's safe to let owners approve these edits?

On the one hand, I suspect this impacts only a small minority of PRs, so
perhaps not a valuable change. On the other hand I think it's a fairly
safe change to make?

Specifically this recategorizes path mapping edits from
`package-meta` (requiring maintainer approval) to
`package-meta-ok` (requiring owner approval).

Co-authored-by: Eli Barzilay <eli@barzilay.org>