Add "please wait for an offer" comments for premature ready-to-merge
Requires the actual merge offer comment to appear, and if it isn't found
then post a message telling the requestor to wait until it does.

Since it relies on the comment, there is a very short timeframe (while
the bot is processing new events, around a minute) where the request is
invalid only because the offer wasn't posted.  The idea is to make
people get used to the idea that they cannot expedite the process by
pre-requesting a merge (which is exactly what leads to instant merge
surprises that nobody likes).

Part of the fix for #216.