New `run` script
There's a bunch of stuff in `commands` and `scripts`, but it's a mess to
update all of them.  Instead, this is a single `run` script that can be
used as a drop-in replacement for `daily`, which can do a dry run, show
various info (including the new `ExtendedPrInfo`) in multiple formats,
and can also operate on a selection of PRs.