Rename `Config Edit` -> `Check Config` and show it for new packages
The information of having this label is useful for new packages too,
since it's added when the new config files don't get auto-blessed by the
file checkers.

OTOH, `Config Edit` is a bad name, since it's only used for the configs
that are not found to be OK.  So something like `Suspicious Config`
would be better --- but it sounds like a potentially malicious content,
so settle for `Check Config`.

Supersedes: #273.

Co-authored-by: Jack Bates <jack@nottheoilrig.com>