Fix re-running of queries in pr-info
Previously, I changed it to re-run the query if `mergeable` is `UNKNOWN`
because it sometimes takes a few seconds for the GH api to update this
information.  Turns out that this should be done *only* for PRs that are
`OPEN`, in other states (`CLOSED` & `MERGED`) it's *expected* to be
`UNKNOWN`.  This broke the removal of project cards for closed/merged
PRs, since it would just spin the 5 retries while getting `UNKNOWN`
every time, eventually treating it as bogus (and also timing out the
gateway).

Should fix #278.