Fix `README.md` links
Looks like `#foo` anchors work when viewing the DT page, for for links
from other pages, the anchors usually fail.  Looking at the actual
rendered HTML, the ids all start with `user-content-`.  So add that, and
do so via a function if they change that in the future.

Also, use the two links for the section on tesing new/existing packages.