Always show the list of files to check
Motivated by the recent change of showing the `Check Config` label also
for new packages, where in such cases the usual place where the
suspicious configs are shown is not reached since there's an earlier
case that shows a different comment for new packages.  Instead of
showing the list of files there too, this just moves the whole thing
into the packages display section (which shows the links and pre-package
approvals).