Remove a bunch of outdated scripts
Many of these are superseded by `run.ts` (showing info, actions,
mutations; dry runs; operate on a specified PR; cleanup columns), and
are no longer needed.  Also shuffled the list of scripts in
`package.json` a bit.

(There are some more things that are not needed, but this is good for
now...)