Ignore nonexistent PR numbers
Make `prFromCheckSuiteEvent` return a new ignore-this value when we
couldn't find a PR and there are `check_suite.pull_requests`.  From
looking at the logs, it seems to me that all of them are "reverse"
PRs (used by forks to pull updates from DT, mostly by a "pull bot").

Keep throwing an error if no PR was found and there are no such reverse
PRs.