Revise the ready-to-merge message to be a bit less celebratory
Since I think that the "great job" part can be confusingly read as an
approving review for people who don't realize that it's not a human who
said it.  I also realized that it's what was making me accompany a PR
blessing with my own comment clarifying that I'm making it possible to
merge "if you're actually done" -- so I added the "whenever you think
it's ready".